requirements.txt
django
python-dotenv
django-extensions
django-celery-beat
celery[redis]
requests
djangorestframework